Refactor AddEventButton and UserEventsList to streamline event handling
* Updated AddEventButton to utilize the new useAddEventFlow hook for improved event creation flow, enhancing code clarity and maintainability.
* Removed unused imports and unnecessary state management from AddEventButton, simplifying the component structure.
* Refactored UserEventsList to integrate useAddEventFlow for triggering event addition, improving consistency across components.
* Introduced a new useAddEventFlow hook to encapsulate the event addition logic, including paywall checks and image selection, enhancing modularity and reusability.
